requirements.txt
Flask==0.12.0
Pillow==4.0.0
